backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.10.0
pydantic==2.5.3
python-dotenv==1.0.0
web3==6.15.1
eth-typing==3.5.0  # Compatible version with web3 6.15.1
httpx==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.2.0
# Monitoring
sentry-sdk[fastapi]==2.9.0
prometheus-client==0.19.0


# Data Science & Monitoring (Consolidated)
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
shap==0.44.0
networkx==3.2.1
psycopg2-binary==2.9.9
redis==5.0.1
pyyaml==6.0.1
a2wsgi==1.10.0
